requirements.txt

# aiogram==3.4.1
# motor==3.6.0
# pymongo==4.10.1
# python-dotenv==1.0.1
# aiohttp==3.11.10
# flask==3.1.0
# apscheduler==3.10.4
# cryptography==44.0.0

# requirements.txt - Fixed versions for Python 3.9-3.12

# Core Bot Framework
aiogram==3.4.1
aiohttp==3.9.3

# Database
motor==3.3.2
pymongo==4.6.1

# Environment & Config
python-dotenv==1.0.1

# Web Server (Phase 2)
flask==3.0.2
werkzeug==3.0.1

# Scheduler
apscheduler==3.10.4
pytz==2024.1

# Security
cryptography==42.0.5

# Additional dependencies
typing-extensions==4.9.0
certifi==2024.2.2

psutil